requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.2
google-cloud-bigquery-datatransfer==3.11.1
google-cloud-bigquery-storage==2.20.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.3
ipython==8.12.2
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pandas-gbq==0.19.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.0
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.15.0